Linear and quadratic probing hash table
Converting the hash table of two indirections to a more compact linear or
quadratic probing table with one indirection.

In the old hash table we had two indirections to get to the data: probe the
corresponding bucket, then traverse the linked list of nodes to find the first
matching node, and then access the data. The indirections are not efficient and
the data structure is not memory-efficient.

The size of the old structures and their overhead per row, depending on the fill
factor of the hash table:

Overhead per entry if 25%: 56 B
Overhead per entry if 33%: 48 B
Overhead per entry if 50%: 40 B
Overhead per entry if 66%: 36 B
Overhead per entry if 75%: 34 B

In the new hash table structure we use linear or quadratic probing and we merge the
buckets with the nodes (that were dense previously). The DuplicateNodes are used for
storing real duplicates, but not entries with same or aliasing hashes.
In this hash table we do not have to store next pointers for the buckets because of
the linear or quadratic probing.

Since in the common case the buckets contain all the information, we have one less
indirection. In addition to the one less indirection, the overall space overhead
is smaller in the common case where the fill factor is higher than 33%:

Overhead per entry if 25%: 64 B
Overhead per entry if 33%: 48 B
Overhead per entry if 50%: 32 B
Overhead per entry if 66%: 24 B
Overhead per entry if 75%: 21 B

Adding a CheckAndResize() function which should be called once before starting
inserting entries from a batch to a hash table, so that inside the entry-inserting
function we don't have to check for resizes. CheckAndResize() will make sure that the
new hash table size will be enough to fit the projected total number of filled buckets.

Merging the common probing functionality of Find() and Insert() in a common function
(Probe()).

All these changes result in improved performance. For example, in a group by query
(select avg(cs) from (select count(*) as cs from lineitem group by l_partkey) l)
with 2M groups in the TPC-H SF10 database got the following:
(TOT=total time, AGG=time in aggregation)
Baseline TOT: 18.8  AGG: 18.7
Linear: TOT: 10.8  AGG: 8.80
Quardatic: TOT: 11.3  AGG: 9.02

Change-Id: If261a6e16d146e270c7f74c525b2356692cf2588
Reviewed-on: http://gerrit.cloudera.org:8080/77
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins