Add simple RecordServiceInputFormat that can read from a HMS table over JDBC
Adds a simple RecordServiceInputFormat that can be used to read from a HMS table.
Currently, it executes one E2E query over JDBC per-input split, and makes uses of
a query plan hint in Impala that will restrict which data is actually read. Can also
update this to use the RecordService API or the mock RecordService service.

Currently, all results are returned as Text with tab-separated field values.
The target db/table name, and column(s) to read are passed in as Strings in the JobConf.

Includes an Example MR job that uses this input format to count the total number of
records in a table. Currently, this is very slow because results are sent on a per-row
basis rather than by rowbatch.

To run the example execute: bin/run-mr-jdbc-job.sh

Change-Id: Iff7ad2c3e69d41c4ecfa92d9ae6f3ee87faf72e1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5814
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>