Add RecordServiceHiveInputFormat that redirects all Hive table scans to the RecordService
A HiveInputFormat that redirects all HDFS reads to the RecordService. This is
done by extracting the Hive query plan and finding the TableScanOperators. From there
we know the table being scanned and the columns being projected. A
RecordServiceRecordReader is then passed to the HiveRecordReader instead of the
table's actual InputFormat.

This can be enabled to happen transparently by setting this class as the
default hive input format. To do this put the following in the hive-site.xml:

<property>
  <name>hive.input.format</name>
  <value>com.cloudera.recordservice.lib.RecordServiceHiveInputFormat</value>
</property>

There are still some issues with this implementation:
- It doesn't handle multiple table scans of different tables.
- Hive short-circuits some statements like SELECT * (it just cats the file).
  These will not go through the RecordService.
- Results may be returned to the Hive client as NULL in some cases since our
  RecordReader is a hack.
- Hive queries that don't specify any columns and multi-table joins don't work

Change-Id: I6e70a2991abc9d5c156dad7f774f661139342739
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5854
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>