Disable concurrency in Hive to allow Hive to run w/o Zookeeper
Hive has a dependency on Zookeeper (queries hang) when locking is enabled.
The HBase mini-cluster started Zookeeper in-process, with the recent
change to disable HBase, Hive is broken. This fixes the problem by
disabling locking in Hive, which we don't needed enabled.

Note that this does cause log spew because Hive still tries to connect
to ZK occasionally, but is a temp workaround until:
a) we find how to run Hive cleanly w/o ZK
b) we start running ZK

Change-Id: I6c392740fa7bf44e150175650ccedcf96ce3c312
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6024
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins