IMPALA-1913: Intel compiler issues with __builtin_prefetch etc.
In the gcc version of prefetch(), hint is only a constant _after_
inlining (assumed to have been successful).  Intel compiler
views things differently, and checks constant-ness _before_ inlining.
This leads to compilation errors with using the other version
of this code with Intel compiler.

One way round this is to use a switch statement to explicitly match
prefetch hint enumerations, and invoke __builtin_prefetch for each
valid value.

Change-Id: Ie6a46cb239e76988a0b9fdf9809514be6f68f82c
Reviewed-on: http://gerrit.cloudera.org:8080/292
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins