IMPALA-1588: HDFS FH Caching - Passing mtime to scan range (Part I)
This patch passes the file modified time from the file descriptor down
to the scan range. The final goal is to perform caching of the file
handle based on the file name and the modified time.

The motivation behind the patch is as follows: The granularity of the
catalog version number is to coarse to be used to invalidate a cache
entry when HDFS file handles are cached. As a consequence, we intend to
use the file path and the mtime as a combined key to identify if a file
needs to be reopened. If the mtime does not match the one of a cached
handle, the handle will be closed and then reopened.

Change-Id: I39af75b38da0bc8cd317f52c4ad951ce1a0d5a0e
Reviewed-on: http://gerrit.cloudera.org:8080/299
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: Internal Jenkins