Improve PATH requests to support wildcards and schema discovery.
The path type request is now more HDFS like. You can say /path/*.avro to read all
avro files or /path/file to just read file. This is a bit different than the hive/impala
view which works on directories but seems very common in spark/mr examples.

It now also supports inferring the file type. This is more of a POC than anything else
and we might not ship this. It currently just looks at the first file in a directory
and figures out the file type from that.

Change-Id: Ic5bd212464d78a25e8189b4fcd5c0de30157efbb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6309
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>