Prevent DOS due to exhausted handler threads.
The current thrift server has a maximum number of handler threads. If this is
exhausted, the client connection just hangs. This is very undesirable. This
patch changes this by rejecting connections if it is using the last handler. This
saves one handler just to reject connections.

We should remove this logic when the server is fixed.

Change-Id: I75949f95f503037315fe317c325aff79a9109074
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6333
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>