Add statestore topics for workers and planners.
This allows us to run daemons that just export one of the services and not
the other. There is more work to use this information that will come in
follow on patches.

This also updates the statestore registration so only the planner services
get the catalog updates.

Automated testing also requires more work so I've just manually verified
using the debug page for now.

Change-Id: Ia9808ff29e152a559dcfaf3c3a150a3626afed81
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6687
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>