initial work for Hive serde scanner
Currently both the FE serde executor and BE scanner are very simple.
For the FE, it simply find rows from the input buffer and return them
to the BE; for the BE, it doesn't do complicated stuff like row/column
boundary handling.

Now if run "select * ..", the first column will be populated with the
whole row data, while the rest of the columns will be filled with NULL.
Also, if run "select" on columns other than the first one, the impalad
will crash.

Change-Id: Iab80776ac74d783e331491564928de7a13ce2ab0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6849
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>