Change ExecNode::Reset() to take a can_free_tuple_data flag.
The Reset() function is used to support executing an exec-node tree multiple
times with the following sequence of calls:
Init()
Prepare()
Open()
GetNext()*
Reset()
Open()
GetNext()*
...
(repeat Reset()/Open()/GetNext() steps)
...
Close()

For now, we only envision multiple executions of an exec-node tree within a
SubplanExecNode (not part of this patch).

This patch changes the ExecNode::Reset() function to accept a can_free_tuple_data
flag which indicates whether it is safe for the node being reset to deallocate
(or otherwise clear and reuse) memory that is backing tuples returned by that node.

Expected uses of the can_free_tuple_data flag:
Reset() should be called on a node with can_free_tuple_data=false as long
as there are in-flight rows that were produced by that node. The goal is to
ensure that the tuple data remains valid until those rows have been processed
all the way up the exec node tree or until they have reached a blocking node
which must deep copy or otherwise acquire ownership of the tuple data.
Once we can guarantee that no such rows are in-flight anymore, Reset()
should be called with can_free_tuple_data=true.

Even if can_free_tuple_data=true, a node implementing Reset() may chose not
to deallocate/reuse memory. Such memory most always be freed in Close(),
regardless of how a node implements Reset(). In the short term, we plan to
have all nodes deallocate/reuse memory in a Reset() with
can_free_tuple_data=true for simplicity.

To demonstrate how Reset() is expected to be implemented and called,
this patch implements BlockingJoinNode::Reset() and CrossJoinNode::Reset().

Change-Id: I3cf62810397d8baa239fd9ffe3d4eb450e41295f
Reviewed-on: http://gerrit.cloudera.org:8080/444
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins