Implement support for SASL DIGEST-MD5 in server.
There are multiple authentication mechanisms that can be used with SASL. Impala
used to support Kerberos and LDAP. The implementations of delegation tokens relies
on DIGEST-MD5 (another standard SASL mechanism). DIGEST-MD5 is a way to send the
password (not plaintext) and have the server just verify it is correct. In the
case of delegation tokens, this is the token itself.

The Impala SASL support is home grown additions to thrift. This is in contrast to
the hadoop implementations where SASL is supported in thrift. Because of this Impala
contains ported (from java) versions of the core transport layers.

This patch includes a dummy test service that will have the same properties as any
thrift service that needs to support delegation tokens. Specifically:
  1. Supports connections using Kerberos and DIGEST-MD5
  2. Has some APIs that require kerberos (the rpc that created delegation tokens) and
     ones that work on either.

This patch is mostly to test the transport changes and includes a Java client that can
speak DIGEST-MD5. The client logic will be the basis of the RecordServicePlanner/Worker
client updates.

The server currently has a hard-coded token it accepts ("admin"). There is additional work
to use the hadoop security APIs to generate tokens, persist them in ZK and verify them.
The current patch should have comments on how that would work.

Change-Id: I024f2889208aa1825a672289cb624c53143801fe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6865
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>