Add breakpad minidump crash handler.
This also adds some utilities scripts to use this. More details on breakpad here:
https://code.google.com/p/google-breakpad/

In summary, breakpad installs crash handler in the application that generates a
minidump when the process crashes. This minidump is much smaller than core dumps
and intended to be sent across the network. This is used, for example, by chrome
to send all crashes back home. This is a natural fit for CM diagnostic bundles
where including core dumps is not feasible. The minidump contains only a fraction
of the information that is normally in a core dump and most useful to let us
know where to look to try to reproduce the bug. It's also much more extensible
and lets us include custom information in the minidump (e.g. list of running
queries).

The minidump supports adding application buffers that are included in the core
dump. I've added a query log ring buffer that stores the last N queries seen
by the service. We can continue to add this as we find more useful diagnostics.

Example of what is in the application buffer:
E0611 16:23:45.847246 24115 minidump.cc:35] Query Log:
Version: 2.2.0-record-service-cdh5-INTERNAL (8a4f9b153e19ba54aeed5093e3594ff649d3e2bb)
Queries: 1
  Timestamp=1434065025162 Idx=1 User='nong' SQL='select * from tpch.nation' Done=false

Change-Id: I0950fb3eaa36f8005f6c31edf4f1413bf5e8987c