More work on Hive serde scanner
This patch adds more functionality to the frontend Hive serde executor
class. First, now we are passing the serde name and properties to the executor
class, for it to initialize the actual serde object. We also use a boolean
array to indicate which columns are needed for the backend. In the
deserialize() method, it now processes the rows into different types of fields,
using the Hive serde object (instead of just echoing the entire rows back.)

It also adds a few unit tests for two serdes: RegexSerDe and OpenCSVSerDe.
The test class is still not complete - more tests need to be added.

On the backend, it moves the row batch processing logic from the DataSourceScanNode
into a new class DataSourceRowConverter, so that this part can be shared by
both the DataSourceScanNode and the HdfsHiveSerDeScanner class.

Change-Id: I44190ae22da202ea6e65fac635856216ada4e5d6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6981
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>