Remove dcheck that kerberos client connections are always internal.
This used to be the case as it was used only for impalad-impalad BE
connections. Impalad BE connections are not intended to be used by
clients.

The RecordServiceScanNode connects to the external planner and worker
services so this DCHECK is not accurate anymore. The code is able to
handle the internal client connection without any issues.

Change-Id: I7f98de5161d8ca21ef00d9ea907cccd7f0e906c7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7003
Reviewed-by: Chao Sun <chao.sun@cloudera.com>
Tested-by: jenkins