Make HdfsHiveSerdeScanner handle row boundaries
This commit adds row boundary handling for HdfsHiveSerdeScanner in the
backend. Since HdfsTextScanner already does sophisticated processing
on column/row boundaries, in this commit we make HdfsHiveSerdeScanner a
subclass of HdfsTextScanner. This enables it to inherit most of the functionalities.
The difference is mainly in WriteFields, in which HdfsHiveSerdeScanner sends
the raw buffer to FE and then processs results back using DataSourceRowConverter.

It also adds a few end-to-end tests for the scanner, which needs to be activated by
another commit to the client repo.

Change-Id: Ied6761eb2ebb08e03a38621a3eeb97981391a58e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7090
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>