CMake updates to build with llvm-trunk.
We have two versions of clang in toolchain, 3.3 and trunk. We use 3.3 for
code generation (where we use the llvm libraries/headers). This patch uses trunk as
the compiler for asan.

Using the 3.3 compiler doesn't work with asan (we need the gcc-toolchain flag
in newer compilers). There's also a lot of benefits of using a new compiler so
this seems like the better path forward. It's found a few bugs that I'll address
in a follow up patch. GCC still proves to be faster so we probably don't want
to switch to clang permanently yet. Trunk llvm is very stable but if we run into
problems, we can snapshot to a newer version.

I can't quite get clang-trunk to compile when asan is disabled due to some errors
when optimizations are disabled with inline asm problems. I'll keep working on it.

Change-Id: I5a2f81e73984df891159fc80dc34af76917947f6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7102
Tested-by: jenkins
Reviewed-by: Alexander Leblang <alex.leblang@cloudera.com>