IMPALA-1929: Avoiding a DCHECK of NULL hash table in spilled right joins
There is an infrequent case of spilled right- and full-outer joins where we
call NextSpilledProbeRowBatch() with a NULL hash table for the spilled
partition and that leads to DCHECK. We have hard time reproducing this DCHECK.
See IMPALA-1929 for more details.

This patch checks the number of rows of the build side. If the build
side has no rows, then it stops processing that partition. Additionally,
it adds more information in the DCHECK, to help us debug in case this problem
is not solved with this patch.

TODO: Try to have a reproduction locally and add a test, probably in
test_spilling.

Change-Id: Id354cc24e93579b03cd40a7d6ff9bc269998bc9f
Reviewed-on: http://gerrit.cloudera.org:8080/523
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins