Set run-custom-cluster-tests.sh to execute with bash
Fix tests/run-custom-cluster-tests.sh to execute with bash rather than
sh so that the bash builtin commands, pushd and popd, will work.

Use #!/usr/bin/env bash so that it will run the bash in the user’s
environment rather than a hardcoded path. This is the standard used
by most of the scripts, though not all follow it.  Some do not depend
on bash only impala-shell.sh requires bash and uses a fixed path.

Change-Id: Iadb2554775fe0099f594b56df3e933ea35ed8cde
Reviewed-on: http://gerrit.cloudera.org:8080/544
Reviewed-by: Casey Ching <casey@cloudera.com>
Tested-by: Internal Jenkins