IMPALA-2160: Fix temp file overallocation
Previously the temporary file manager would extend a temporary file to
<current length> + 2 * <write_size> bytes. This behaviour appears
incorrect and does not match the method comment. This commit changes the
behavior so that it allocates <current length> + <write_size> bytes.

Simplify the class so that it doesn't track current_offset_ and
current_size_ separately - they are now the same.

The WriteError unit test depended (erroneously) on the old behavior so
that has to be updated as a consequence of the change. We now have two
separate tests for errors allocating backing temporary storage for blocks
and write errors for already-allocated blocks.

Change-Id: Ifecf030cb2648e7c52c94aff15948cec59c37477
Reviewed-on: http://gerrit.cloudera.org:8080/581
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins