Fix race in token expire thread which causes the daemon to exit.
Our daemon starts up a thread which periodically checks for expired tokens.
This is racy if the token is removed by someone else. The thread will hit
an exception which is uncaught. As commented in this code, the way exceptions
are handled in this code makes these kind of errors harder to avoid.

Change-Id: If9d29dc749822902f5374682b8e9b2d0ba138c56
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7279
Reviewed-by: Chao Sun <chao.sun@cloudera.com>
Tested-by: jenkins