Make sure children are created before ExecNode::Init() is called on parent
SubplanNode::Init() expects that its children are already created,
which is not the case before this patch. We could move the work done
in SubplanNode::Init() to SubplanNode::Prepare(), but it seems like a
reasonable invariant of Init() to already have access to Init()'d
children. This patch achieves this invariant by moving the
ExecNode::Init() call from ExecNode::CreateNode() to
ExecNode::CreateTreeHelper().

Change-Id: Ia226d0d4796b96a72bfd9b3be56aa619423e84cf
Reviewed-on: http://gerrit.cloudera.org:8080/593
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins