Plumb zookeeper membership in BE.
We need the membership in the BE for the debug webpage/task assignment, etc so this
adds the final step to not need statestore for membership.

We JNI the updates from the FE on zookeeper updates. When we remove the statestore
dependency completely, I'll remove the relevant code in the BE as well.

Change-Id: I41e2b5b5070fc303806fd729f14fef050f9a6f93
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7445
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>