Show Scan Process for RecordService
At the moment the Scan Process in the debug page only shows
the number of processed splits versus total number of splits, and in
the case of RecordService, this will always be "0/1 (0%)" until the very
end. This changes it to be the number of bytes processed versus
the total number of bytes, so user can track the task progress in the
page and see how much of the task has been finished.

Change-Id: I811467dd7c5a32ddb566261da671d6523e5c25ab
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7507
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>