Fix performance regression from computing number of nodes in HdfsScanNode.
To handle on local ranges (e.g. S3), HdfsScanNode walks all the ranges to compute
the number of nodes this query will run on. If there are a lot of ranges, this can
slow down planning in a noticeable way. This patches adds an optimization so that
if we've found a range for each node in the cluster we stop looking at the remaining
ranges. This should have the identical behavior without the optimization.

Change-Id: I5f12dd0cdee3c5f20b20861241394bb4b505c56a