IMPALA-2178: fix Expr::ComputeResultsLayout() logic
I am honestly not sure what the original code handling CHAR was
supposed to do, but regardless of the intent I don't think it was
right. This patch rewrites the ComputeResultsLayout() logic to change
the handling of CHAR to reflect my understanding of how the compiler
would layout the memory, as well make the rest of the logic easier to
understand.

Change-Id: I4dfc5968c35dc9fc81bb87926a04410d335205b6
Reviewed-on: http://gerrit.cloudera.org:8080/617
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins