Update csd to have planner and worker only roles.
Previously we only had the role of running both the planner and worker. IMO, ideally
for each daemon, we should get to pick if it starts either one or both but that is not
really the CM model.

We can revisit this but I think this deployment flexibility is good for now. I've set
up the 5 node to run 1 planner and 4 workers and the 3 node to run 3 planners and workers.

Change-Id: Ie9ce04047b9d00e934e1ec9a256ab433f387d35e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7825
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: Nong Li <nong@cloudera.com>