Fix how we handle worker service ports.
Previously the planner just assumed the worker port from a planner side config. This
patch updates it so that the worker membership (already maintained by the planner),
contains the port information as well. This was actually already maintained by the
planner and just unused.

This patch also adds logic to ignore local HDFS blocks if there is no worker running
on that node.

Change-Id: I91dd819dfc3b4d5a9c77769ef17ed2fa09ba0227
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7832
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>
Reviewed-by: Nong Li <nong@cloudera.com>