Updates to resolve localhost -> IP address in PlanRequest.
This will allow us to run tests against a remote dev cluster, such as having
the client tests run on OSX to a dev cluster on my box. The configuration of
the dev cluster is a little unusual (compared to CM cluster) because the hostname
is localhost. This is required for some of the other services we depend on (HDFS)
to simulate locality matching. This doesn't work if the planner is returning these
as the address of the worker though.

With this change, I can run the tests from a remote client.

Change-Id: I636d57b8bd2379e86ecd1ce0734cf637bca6892f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7875
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>