IMPALA-2100: Exclude explain header from expected results of test_partitioning.py.
HDFS acknowledges writes when the first replica is written.
As a result, the estimated memory requirements for an Impala
query may vary depending on how many replicas existed at the
time of table loading. This racey behavior caused a few tests
to sometimes fail due to different actual and expected memory
requirements.

The fix is to exclude the explain header from the expected results.

Change-Id: Ifb13de937a104a48960d35745df521de66596837
Reviewed-on: http://gerrit.cloudera.org:8080/762
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins