Enable strict read/write on Thrift server
Currently without strict read, when a secure client
trying to connect to a non-secure server, it will eventually timeout
This is because a bad cast on the thrift thread, which causes it to die.
With this enabled, it will throw exception instead (and will not die),
so the client will be rejected quickly.

Change-Id: Iabc4f2921b063a64c3456bde2e0426ddd9bab87f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7911
Tested-by: jenkins
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>