Implement support to combine tasks in PlanRequest.
Having a very large number of tasks is not good for the cluster. This results in
very RPC response from the planner (1GB for 750K tasks) and causes all kinds of
problems for the other parts of the stack. It is also inefficient to launch so many
tasks due to scheduling overhead, task overhead, etc.

This patch adds the ability to specify a maximum number of tasks PlanRequest() should
generate. If the number of scan ranges is larger, it will collapse them, trying to
balance locality and even task size. This patch includes a simple heuristic based
greedy solution. This can be swapped out in the future with a more accurate solution.

This is very hard to test, especially on the local dev environment where locality
is fabricated. I did try this on the 1 trillion record data set and it appears to
work. Planning times go down (from 35->15) and job time goes down 4x.

Change-Id: Ia996880dfc504c2d51d1cb0f2e2772774ba9e60a