Fix QueryStateRecord, when profile is NULL
Currently when we instantiate a QueryStateRecord, we assume
that if it's RS and has a coordinator, then it must have a
profile for HDFS scan node. However, this might not be true all
the time, due to the query in a cancellation path, etc.

This removes the NULL check, and makes the code to be more defensive.

Change-Id: I182b85420ef47942529c24096b9e32ff1cf0e2f0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8041
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>