IMPALA-2358: Fix join ordering of relative collection table refs inside a subplan.
The bug was that we were dropping a relative collection table ref during plan
generation. As a result, we were not computing the memory layout for the item
tuple descriptor of a collection-typed slot, which lead to a crash in the BE
when trying to populate that slot with an invalid offset during a scan.

Change-Id: I229b04ac581305ce78bc67b094fe7220881bbe46
Reviewed-on: http://gerrit.cloudera.org:8080/883
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins