Fix path request when Sentry is enabled
This is a (temporary) fix of path request when Sentry is enabled.

Currently, the username is not populated when creating the temp table.
With this, it first checks if the user has privilege to the path,
then process the temp table on behalf of the particular user that's
running the recordserviced process. For the latter, we can't use the
current user name because this would enable users to access table
created by others.

Change-Id: Ib632327d221c8b483fdfcc5369cb2b55fbfad179
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8584
Tested-by: jenkins
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Reviewed-by: Chao Sun <chao.sun@cloudera.com>